added graphing
