+++ logs/backend.log
@@ -4802,3 +4802,87 @@
 [2023-09-28 19:31:24,551][worker-0][INFO] Initialized Seeds
 [2023-09-28 19:31:24,552][worker-0][INFO] Initialized CORS
 [2023-09-28 19:31:24,553][worker-0][INFO] App Initialization is finished successfully
+[2023-10-01 11:31:11,701][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-01 11:31:11,726][worker-0][INFO] Initialized Seeds
+[2023-10-01 11:31:11,726][worker-0][INFO] Initialized CORS
+[2023-10-01 11:31:11,726][worker-0][INFO] App Initialization is finished successfully
+[2023-10-01 11:41:39,935][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-01 11:41:39,951][worker-0][INFO] Initialized Seeds
+[2023-10-01 11:41:39,951][worker-0][INFO] Initialized CORS
+[2023-10-01 11:41:39,951][worker-0][INFO] App Initialization is finished successfully
+[2023-10-01 11:44:38,420][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-01 11:44:38,436][worker-0][INFO] Initialized Seeds
+[2023-10-01 11:44:38,436][worker-0][INFO] Initialized CORS
+[2023-10-01 11:44:38,436][worker-0][INFO] App Initialization is finished successfully
+[2023-10-01 11:49:08,799][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-01 11:49:08,818][worker-0][INFO] Initialized Seeds
+[2023-10-01 11:49:08,818][worker-0][INFO] Initialized CORS
+[2023-10-01 11:49:08,818][worker-0][INFO] App Initialization is finished successfully
+[2023-10-01 11:49:47,967][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-01 11:49:47,987][worker-0][INFO] Initialized Seeds
+[2023-10-01 11:49:47,988][worker-0][INFO] Initialized CORS
+[2023-10-01 11:49:47,988][worker-0][INFO] App Initialization is finished successfully
+[2023-10-05 21:55:48,279][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-05 21:55:48,339][worker-0][INFO] Initialized Seeds
+[2023-10-05 21:55:48,340][worker-0][INFO] Initialized CORS
+[2023-10-05 21:55:48,340][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:32:04,621][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:32:04,659][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:32:04,660][worker-0][INFO] Initialized CORS
+[2023-10-06 00:32:04,660][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:33:51,177][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:33:51,213][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:33:51,214][worker-0][INFO] Initialized CORS
+[2023-10-06 00:33:51,214][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:35:10,696][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:35:10,727][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:35:10,728][worker-0][INFO] Initialized CORS
+[2023-10-06 00:35:10,728][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:37:16,707][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:37:16,747][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:37:16,748][worker-0][INFO] Initialized CORS
+[2023-10-06 00:37:16,748][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:39:19,510][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:39:19,554][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:39:19,554][worker-0][INFO] Initialized CORS
+[2023-10-06 00:39:19,555][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:40:56,817][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:40:56,857][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:40:56,857][worker-0][INFO] Initialized CORS
+[2023-10-06 00:40:56,858][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:43:10,673][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:43:10,707][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:43:10,707][worker-0][INFO] Initialized CORS
+[2023-10-06 00:43:10,707][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:44:14,359][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:44:14,435][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:44:14,436][worker-0][INFO] Initialized CORS
+[2023-10-06 00:44:14,437][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:45:47,412][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:45:47,451][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:45:47,451][worker-0][INFO] Initialized CORS
+[2023-10-06 00:45:47,452][worker-0][INFO] App Initialization is finished successfully
+[2023-10-06 00:47:35,216][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-06 00:47:35,257][worker-0][INFO] Initialized Seeds
+[2023-10-06 00:47:35,258][worker-0][INFO] Initialized CORS
+[2023-10-06 00:47:35,258][worker-0][INFO] App Initialization is finished successfully
+[2023-10-12 14:14:25,655][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-12 14:14:25,673][worker-0][INFO] Initialized Seeds
+[2023-10-12 14:14:25,673][worker-0][INFO] Initialized CORS
+[2023-10-12 14:14:25,673][worker-0][INFO] App Initialization is finished successfully
+[2023-10-12 14:26:16,116][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-12 14:26:16,132][worker-0][INFO] Initialized Seeds
+[2023-10-12 14:26:16,132][worker-0][INFO] Initialized CORS
+[2023-10-12 14:26:16,132][worker-0][INFO] App Initialization is finished successfully
+[2023-10-12 14:28:54,481][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-12 14:28:54,500][worker-0][INFO] Initialized Seeds
+[2023-10-12 14:28:54,500][worker-0][INFO] Initialized CORS
+[2023-10-12 14:28:54,500][worker-0][INFO] App Initialization is finished successfully
+[2023-10-12 14:30:37,865][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-12 14:30:37,882][worker-0][INFO] Initialized Seeds
+[2023-10-12 14:30:37,882][worker-0][INFO] Initialized CORS
+[2023-10-12 14:30:37,883][worker-0][INFO] App Initialization is finished successfully
+[2023-10-12 14:31:30,096][worker-0][INFO] Initialized logger with level 10 for worker 0
+[2023-10-12 14:31:30,112][worker-0][INFO] Initialized Seeds
+[2023-10-12 14:31:30,113][worker-0][INFO] Initialized CORS
+[2023-10-12 14:31:30,113][worker-0][INFO] App Initialization is finished successfully
