+++ portal/handlers/differentiation_handler.py
@@ -4,8 +4,13 @@ import sympy as sp
 def calculate_differentiation(data):
     print(data[0])
     expr_str = data[0]
-    var_str = data[1]
-    var = sp.symbols(var_str)
+    if len(data) > 1:
+        var_str = data[1]
+        var = sp.symbols(var_str)
+    else:
+        var = sp.symbols("x")
+    # var_str = data[1]
+    # var = sp.symbols(var_str)
 
     # Parse the expression
     expr = sp.sympify(expr_str)
