+++ portal/templates/graphing.html
@@ -1,0 +1,121 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>Calculus</title>
+
+    <link rel="stylesheet" type="text/css" href="/uncg_math/static/Geometry_files/mystyle.css">
+    <script src="/uncg_math/static/Geometry_files/geometry.js"></script>
+</head>
+
+<body>
+    <div class="example" style="height: 100% !important;">
+        <div id="nav" style="clear:both;">
+            <h1 style="color: white;text-align: center;"> Graphing</h1>
+
+            <ul style="margin-top: -20px;">
+                <li style="padding-left: 179px;"><a href="/uncg_math">Home</a></li>
+<!--                <li style="padding-left: 179px;"><a href="/uncg_math/algebra">Back</a></li>-->
+            </ul>
+        </div>
+
+        <!-- Section for Limits -->
+        <!-- Section for Limits -->
+<h2>Graphing</h2>
+<FORM>
+    <TABLE border=0 width=100% bgcolor=moccasin>
+        <tr>
+            <td> <br> </td>
+        </tr>
+
+
+        <TR>
+            <TD COLSPAN=7 ALIGN=LEFT VALIGN=MIDDLE>
+                <span>Please enter the function:</span><br>
+                <INPUT TYPE=TEXT NAME="limit_input" ID="limit_input" style="padding: 12px; width: 250px;">
+                <INPUT TYPE="button" class="widebutton" VALUE="Calculate" onclick="get_limit()">
+            </TD>
+        </TR>
+        <tr>
+            <TD ALIGN=CENTER>Result :</TD>
+            <TD><span id="limit_result" style="font-size: small;">-------------------</span></TD>
+        </tr>
+    </TABLE>
+</FORM>
+
+        <!-- Other sections (Differentiation, Integration, Series) can be added in the same format -->
+
+    </div>
+</body>
+
+<script src="http://code.jquery.com/jquery-1.9.1.js"></script>
+    <script src="http://code.jquery.com/ui/1.11.0/jquery-ui.js"></script>
+
+<script>
+function get_limit1(){
+        eqn = document.getElementById("limit_input").value
+
+
+        data_dict = {'eqn':eqn}
+
+
+        $.ajax({
+                url: '/uncg_math/graphing',
+                data: JSON.stringify(data_dict),
+                type: 'POST',
+                contentType: "application/json",
+                dataType: 'json',
+                success: function (data2) {
+                    document.getElementById("limit_result").innerText = data2['result'];
+                   return true;
+                },
+                error: function (data2) {
+                    console.log(data2);
+                    return false;
+                }
+            });
+
+
+    }
+
+function get_limit() {
+    eqn = document.getElementById("limit_input").value
+    data_dict = {'eqn': eqn}
+
+    $.ajax({
+        url: '/uncg_math/graphing',
+        data: JSON.stringify(data_dict),
+        type: 'POST',
+        contentType: "application/json",
+        dataType: 'json',
+        success: function (data2) {
+            var result = data2['result'];
+
+            // Format and display the values from the result object
+            var domain = result['Domain'].join(', '); // Join array elements into a string
+            var range = result['Range'][0] + " to " + result['Range'][1];
+            var period = result['Period'] || "N/A";
+            var maxima = result['Maxima'] || "N/A";
+            var minima = result['Minima'] || "N/A";
+            var symmetric = result['Symmetric'] ? "Yes" : "No";
+
+            var outputText = "Domain: " + domain + "\n" +
+                            "Range: " + range + "\n" +
+                            "Period: " + period + "\n" +
+                            "Maxima: " + maxima + "\n" +
+                            "Minima: " + minima + "\n" +
+                            "Symmetric: " + symmetric;
+
+            document.getElementById("limit_result").innerText = outputText;
+        },
+        error: function (data2) {
+            console.log(data2);
+        }
+    });
+}
+
+
+ </script>
+
+</html>
