+++ portal/templates/home.html
@@ -218,6 +218,9 @@ the content or design. Without your permission, your email
     <td>
       <h3><a href="/uncg_math/trigonometry" style="width: 200px; height: 20px">Trigonometry</a></h3>
     </td>
+       <td>
+      <h3><a href="/uncg_math/graphing" style="width: 200px; height: 20px">Graphing</a></h3>
+    </td>
 <!--    <td>-->
 <!--      <h3><a href="/uncg_math/arithmetic_expressions" style="width: 200px; height: 20px">Arithmetic Exp </a></h3>-->
 <!--    </td>-->
